+++ CategoryMicro/pom.xml
@@ -85,6 +85,11 @@
             <artifactId>grizlibrary</artifactId>
             <version>0.0.5-SNAPSHOT</version>
         </dependency>
+        <dependency>
+            <groupId>org.springframework.boot</groupId>
+            <artifactId>spring-boot-configuration-processor</artifactId>
+            <version>2.0.3.RELEASE</version>
+        </dependency>
     </dependencies>
